+++ CHANGES.md
@@ -21,6 +21,7 @@
   - ➕ `charAt` / `char_at`
   - ➕ `charCodeAt` / `char_code_at`
   - ➕ `codePointAt` / `code_point_at`
+  - ➕ `concat` => `jsConcat` / `js_concat`
   - ➕ `fromCharCode` / `from_char_code`
   - ➕ `fromCodePoint` / `from_code_point`
   - ➕ `includes`